# HDDS-8822. [S3G] Improve list performance in LEGACY/OBS bucket