--- a/hadoop-ozone/s3gateway/src/test/java/org/apache/hadoop/ozone/s3/endpoint/TestBucketList.java
+++ b/hadoop-ozone/s3gateway/src/test/java/org/apache/hadoop/ozone/s3/endpoint/TestBucketList.java
@@ -36,6 +36,8 @@ import org.junit.Test;
 
 /**
  * Testing basic object list browsing.
+ * Note: delimiter with '/' will call shallow list logic,
+ * just list immediate subdir of prefix.
  */
 public class TestBucketList {