@whbing 2023-07-22
Sorry for the late follow-up. I added the test and method doc for `listKeys`, and also tested the s3 list for legacy and obs buckets in the test environment. @captainzmc   please help triger ci, thanks !