@whbing 2023-08-10
@captainzmc Thanks for review. I added a param `ozone-s3g.list-keys.shallow.enabled`  to compare the efficiency before and after using this feature. The time taken for both the `true` and `false` settings corresponds with the initial description provided on this page.